cptools.SessionAuth: fix login_screen page handler
The login_screen page handler erroneously returned the formatted HTML page,
which was simply interpreted as a boolean that controls whether the call
to the controller function was suppressed or not.

Set the response.body instead and return a literal True that suppresses the
call to the page handler.

Signed-off-by: Balazs Scheidler <bazsi@balabit.hu>

--HG--
branch : fix-session-auth-login-screen